feat: implement middleware for authentication, logging, and tenant resolution
